% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados.R
\name{regressao918}
\alias{regressao918}
\title{Regressão Linear}
\usage{
regressao918(Y, X, db)
}
\arguments{
\item{Y}{Nome, entre aspas duplas, da variável resposta, identico ao nome da coluna do banco de dados utilizado}

\item{X}{Vetor de nomes, entre aspas duplas, das variáveis preditoras, identico aos nomes das colunas do banco de dados utilizado}

\item{db}{Nome do banco de dados}
}
\value{
Lista com os coeficientes estimados, os valores preditos, resíduos e gráfico (Valores Preditos vs. Valores Observados)
}
\description{
Função que recebe um data frame e, pelo menos, o nome de duas colunas deste.
Aplica uma regressão linear entre as variáveis especificadas.
}
\examples{

regressao("sim_Y", "sim_X", sim_dados)

}
